Velocity.getDeg()の戻り値をint型に変更

+++ src/casestudy/Velocity.java
@@ -24,7 +24,7 @@ public class Velocity {
 		return vel_;
 	}
 
-	public double getDeg() {
+	public int getDeg() {
 		return deg_;
 	}
